Move current_blind to game vars
